pyproject.toml
[project]
name = "goboscript-discord"
version = "0.1.0"
description = "The goboscript discord bot."
readme = "README.md"
requires-python = ">=3.13.5"
dependencies = [
    "anthropic>=0.54.0",
    "discord>=2.3.2",
    "httpx>=0.28.1",
    "langchain>=0.3.25",
    "numpy>=2.3.0",
    "python-dotenv>=1.1.0",
    "redisvl>=0.7.0",
    "rich-click>=1.8.9",
    "sb2gs",
    "voyageai>=0.3.2",
]

[tool.pyright]
reportUnnecessaryTypeIgnoreComment = true
reportUnknownVariableType = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportMissingModuleSource = false
reportMissingTypeStubs = false
reportWildcardImportFromLibrary = false
reportPrivateUsage = false
reportPrivateImportUsage = false
reportAny = false
reportExplicitAny = false
reportUnusedCallResult = false

[tool.ruff.lint]
select = ["ALL"]
extend-safe-fixes = ["ALL"]
unfixable = ["F841"]
ignore = ["D", "COM", "ANN401", "S101"]

[tool.ruff.lint.isort]
required-imports = ["from __future__ import annotations"]

[tool.uv.sources]
sb2gs = { path = "../sb2gs" }
